extract computed_property, relying on TS to resolve circular type deps
